backend-users-model-custom-user refactor: add file forms and change naming
