+++ backend/spanish_with_anna/users/admin.py
@@ -1,10 +1,8 @@
-from django import forms
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
 from django.contrib.auth.models import Group
-from django.contrib.auth.forms import ReadOnlyPasswordHashField
-from django.core.exceptions import ValidationError
 
+from users.forms import CustomUserChangeForm, CustomUserCreationForm
 from users.models import CustomUser, Feedback
 
 
@@ -18,64 +16,6 @@ class AdminFeedback(admin.ModelAdmin):
     list_display = ('name', 'email', 'phone', 'time_create', 'user', 'message')
     list_filter = ('email', 'time_create', 'user')
     search_fields = ('email', 'message')
-
-
-class CustomUserCreationForm(forms.ModelForm):
-    """
-    Форма для создания пользователя.
-    """
-
-    password1 = forms.CharField(
-        label='Пароль',
-        widget=forms.PasswordInput,
-    )
-    password2 = forms.CharField(
-        label='Подтверждение пароля',
-        widget=forms.PasswordInput,
-    )
-
-    class Meta:
-        model = CustomUser
-        fields = ('email', 'name', 'role', 'phone',)
-
-    def clean_password2(self):
-        """
-        Функция проверяет, что введенные пароли совпадают.
-        """
-        password1 = self.cleaned_data.get('password1')
-        password2 = self.cleaned_data.get('password2')
-        if password1 and password2 and password1 != password2:
-            raise ValidationError("Указанные пароли нес совпадают.")
-
-        return password2
-
-    def save(self, commit=True):
-        """
-        Функция создает пользователя и сохраняет пароль в хешированном виде.
-        """
-        user = super().save(commit=False)
-        user.set_password(self.cleaned_data['password1'])
-        if commit:
-            user.save()
-
-        return user
-
-
-class CustomUserChangeForm(forms.ModelForm):
-    """
-    Форма для обновления пользователей. Включает все поля пользователя.
-    """
-
-    password = ReadOnlyPasswordHashField(label=(" Пароль"), help_text=(
-        "Просмотреть пароль пользователя невозможно. Для изменения "
-        "пароля используйте <a href=\"../password/\">эту форму</a>."))
-
-    class Meta:
-        model = CustomUser
-        fields = (
-            'name', 'email', 'password', 'phone', 'role', 'courses',
-            'is_active', 'is_staff', 'is_superuser',
-        )
 
 
 class CustomUserAdmin(BaseUserAdmin):
@@ -112,7 +52,8 @@ class CustomUserAdmin(BaseUserAdmin):
     add_fieldsets = (
         (None, {
             'fields': (
-                'email', 'name', 'phone', 'role', 'password1', 'password2',
+                'email', 'name', 'phone', 'role', 'password',
+                'password_confirmation',
             ),
         }),
         (('Permissions'), {
